fix: clear validation debt before rebuilding SplitShot
Before:
- Repository-wide Ruff validation reported 249 findings across runtime, audit, and test code.
- Two analysis tests still asserted superseded cross-stage inheritance instead of application defaults.

Now:
- Ruff lint and formatting pass across the entire repository.
- Stage creation and first media import tests enforce application-default isolation.
- Python utility boundaries, exception types, subprocess intent, imports, and executable script modes are explicit and clean.

Why:
- All known validation failures must be resolved before producing another application build.
- Linked to: manager request to fix every outstanding issue before rebuilding.

How:
- Applied Ruff safe and reviewed unsafe fixes, then resolved remaining findings explicitly.
- Updated stale stage-default tests to the approved application-only settings contract.
- Re-ran browser interaction, full canonical isolation, packaging, signing, and packaged launch verification.

Proof:
- 79/79 canonical one-by-one test targets passed.
- Browser interaction audit passed.
- uvx ruff check . --no-cache passed.
- uvx ruff format --check . passed.
- JavaScript syntax checks and git diff --check passed.
- macOS arm64 build, bundle verification, codesign verification, and packaged smoke launch passed.

+++ src/splitshot/presentation/popups.py
@@ -8,7 +8,6 @@ from splitshot.scoring.logic import (
 )
 from splitshot.timeline.model import sort_shots
 
-
 POPUP_BUBBLE_QUADRANT_POINTS = {
     "top_left": (0.125, 0.125),
     "top_middle": (0.5, 0.125),
@@ -75,7 +74,7 @@ def popup_bubble_time_ms(project: Project, popup: object) -> int:
         if shot is not None:
             return shot.time_ms
     try:
-        return max(0, int(round(float(_field(popup, "time_ms", 0) or 0))))
+        return max(0, round(float(_field(popup, "time_ms", 0) or 0)))
     except (TypeError, ValueError):
         return 0
 
@@ -83,7 +82,7 @@ def popup_bubble_time_ms(project: Project, popup: object) -> int:
 def popup_bubble_visible_window(project: Project, popup: object) -> tuple[int, int]:
     start_ms = popup_bubble_time_ms(project, popup)
     try:
-        duration_ms = max(1, int(round(float(_field(popup, "duration_ms", 1000) or 1000))))
+        duration_ms = max(1, round(float(_field(popup, "duration_ms", 1000) or 1000)))
     except (TypeError, ValueError):
         duration_ms = 1000
     return start_ms, start_ms + duration_ms
@@ -149,7 +148,7 @@ def popup_bubble_motion_path(popup: object) -> list[tuple[int, float, float, str
     for item in raw_path:
         try:
             offset_ms = max(
-                0, int(round(float(_field(item, "offset_ms", _field(item, "time_ms", 0)) or 0)))
+                0, round(float(_field(item, "offset_ms", _field(item, "time_ms", 0)) or 0))
             )
         except (TypeError, ValueError):
             offset_ms = 0
